# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Platform API",
    "description": "Enterprise Healthcare Platform Foundation API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/interoperability/patients/{patient_id}/identifiers": {
      "post": {
        "tags": [
          "ABDM & Interoperability"
        ],
        "summary": "Link ABHA or External Identifier to Patient",
        "operationId": "link_identifier_api_v1_interoperability_patients__patient_id__identifiers_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientIdentifierCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientIdentifierResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ABDM & Interoperability"
        ],
        "summary": "List Linked External Identifiers for Patient",
        "operationId": "list_identifiers_api_v1_interoperability_patients__patient_id__identifiers_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientIdentifierResponse"
                  },
                  "title": "Response List Identifiers Api V1 Interoperability Patients  Patient Id  Identifiers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interoperability/consents": {
      "post": {
        "tags": [
          "ABDM & Interoperability"
        ],
        "summary": "Register Patient Consent Artefact",
        "operationId": "create_consent_api_v1_interoperability_consents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/interoperability/consents/{consent_id}/revoke": {
      "post": {
        "tags": [
          "ABDM & Interoperability"
        ],
        "summary": "Revoke Patient Consent Artefact",
        "operationId": "revoke_consent_api_v1_interoperability_consents__consent_id__revoke_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interoperability/abdm/verify-abha/{abha_number}": {
      "get": {
        "tags": [
          "ABDM & Interoperability"
        ],
        "summary": "Simulate ABHA Number Verification via ABDM Gateway",
        "operationId": "verify_abha_number_api_v1_interoperability_abdm_verify_abha__abha_number__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "abha_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Abha Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Verify Abha Number Api V1 Interoperability Abdm Verify Abha  Abha Number  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/encounters/{encounter_id}/prescriptions": {
      "post": {
        "tags": [
          "Prescriptions"
        ],
        "summary": "Create and issue prescription for an encounter",
        "description": "Licensed prescriber generates a prescription with validated medication line items.",
        "operationId": "create_prescription_api_v1_encounters__encounter_id__prescriptions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Prescriptions"
        ],
        "summary": "Get prescriptions created during encounter",
        "description": "Retrieve all prescriptions written during a clinical encounter.",
        "operationId": "get_encounter_prescriptions_api_v1_encounters__encounter_id__prescriptions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionResponse"
                  },
                  "title": "Response Get Encounter Prescriptions Api V1 Encounters  Encounter Id  Prescriptions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}": {
      "get": {
        "tags": [
          "Prescriptions"
        ],
        "summary": "Get prescription details",
        "description": "Retrieve full details and items of a specific prescription.",
        "operationId": "get_prescription_api_v1_prescriptions__prescription_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/prescriptions": {
      "get": {
        "tags": [
          "Prescriptions"
        ],
        "summary": "Get longitudinal prescription history for patient",
        "description": "Retrieve paginated prescription history for a patient.",
        "operationId": "get_patient_prescriptions_api_v1_patients__patient_id__prescriptions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/issue": {
      "post": {
        "tags": [
          "Prescriptions"
        ],
        "summary": "Issue prescription",
        "description": "Formally issue a draft prescription.",
        "operationId": "issue_prescription_api_v1_prescriptions__prescription_id__issue_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prescriptions/{prescription_id}/cancel": {
      "post": {
        "tags": [
          "Prescriptions"
        ],
        "summary": "Cancel prescription with audit reason",
        "description": "Cancel a prescription and record reason and clinician provenance.",
        "operationId": "cancel_prescription_api_v1_prescriptions__prescription_id__cancel_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prescription Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentCreateRequest": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "purpose": {
            "$ref": "#/components/schemas/ConsentPurpose",
            "default": "CARE_MANAGEMENT"
          },
          "scope": {
            "$ref": "#/components/schemas/ConsentScope",
            "default": "ALL"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "patient_id"
        ],
        "title": "ConsentCreateRequest"
      },
...
      "ConsentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "consent_artefact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Artefact Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "granted_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granted By"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "purpose",
          "scope",
          "status",
          "granted_at"
        ],
        "title": "ConsentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentifierType": {
        "type": "string",
        "enum": [
          "ABHA_NUMBER",
          "ABHA_ADDRESS",
          "NATIONAL_ID",
          "PASSPORT",
          "DRIVING_LICENSE"
        ],
        "title": "IdentifierType"
      },
...
      "PatientIdentifierCreate": {
        "properties": {
          "system": {
            "type": "string",
            "title": "System",
            "description": "Authority or namespace URI",
            "default": "https://healthid.abdm.gov.in"
          },
          "value": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Value",
            "description": "Identifier value (e.g. ABHA number or address)"
          },
          "identifier_type": {
            "$ref": "#/components/schemas/IdentifierType",
            "default": "ABHA_NUMBER"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "PatientIdentifierCreate"
      },
      "PatientIdentifierResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "system": {
            "type": "string",
            "title": "System"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "identifier_type": {
            "type": "string",
            "title": "Identifier Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "system",
          "value",
          "identifier_type",
          "status",
          "created_at"
        ],
        "title": "PatientIdentifierResponse"
      },
...
      "PrescriptionCancelRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "PrescriptionCancelRequest"
      },
      "PrescriptionCreate": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PrescriptionCreate"
      },
...
      "PrescriptionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PrescriptionListResponse"
      },
      "PrescriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "encounter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Encounter Id"
          },
          "prescriber_id": {
            "type": "integer",
            "title": "Prescriber Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "status": {
            "$ref": "#/components/schemas/PrescriptionStatusEnum"
          },
          "prescribed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Prescribed At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "issued_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued By"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "cancelled_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled By"
          },
          "cancellation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Reason"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemResponse"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "encounter_id",
          "prescriber_id",
          "facility_id",
          "status",
          "prescribed_at",
          "created_at"
        ],
        "title": "PrescriptionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}